modified the folder structure for waste water reduction calculator, because it's been moved from utilities (general) to waste-water.
