+++ src/app/calculator/calculator-routing/general.routing.ts
@@ -7,7 +7,7 @@ import { NaturalGasReductionComponent } from "../utilities/natural-gas-reduction
 import { PowerFactorCorrectionComponent } from "../utilities/power-factor-correction/power-factor-correction.component";
 import { PreAssessmentComponent } from "../utilities/pre-assessment/pre-assessment.component";
 import { UnitConverterComponent } from "../utilities/unit-converter/unit-converter.component";
-import { WaterReductionComponent } from "../utilities/water-reduction/water-reduction.component";
+import { WaterReductionComponent } from "../waste-water/water-reduction/water-reduction.component";
 import { UtilitiesListComponent } from "../utilities/utilities-list/utilities-list.component";
 
 export const generalRoutes: Routes = [
