+++ src/app/calculator/utilities/utilities.module.ts
@@ -7,7 +7,6 @@ import { PowerFactorCorrectionModule } from './power-factor-correction/power-fac
 import { Co2SavingsModule } from './co2-savings/co2-savings.module';
 import { ElectricityReductionModule } from './electricity-reduction/electricity-reduction.module';
 import { NaturalGasReductionModule } from './natural-gas-reduction/natural-gas-reduction.module';
-import { WaterReductionModule } from './water-reduction/water-reduction.module';
 import { CashFlowModule } from './cash-flow/cash-flow.module';
 import { UnitConverterModule } from './unit-converter/unit-converter.module';
 import { UtilitiesListComponent } from './utilities-list/utilities-list.component';
@@ -31,7 +30,6 @@ import { FanPsychrometricModule } from './fan-psychrometric/fan-psychrometric.mo
         Co2SavingsModule,
         ElectricityReductionModule,
         NaturalGasReductionModule,
-        WaterReductionModule,
         CashFlowModule,
         UnitConverterModule,
         RouterModule,
