+++ src/app/calculator/waste-water/waste-water.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { WasteWaterListComponent } from './waste-water-list/waste-water-list.component';
 import { RouterModule } from '@angular/router';
 import { O2UtilizationRateModule } from './o2-utilization-rate/o2-utilization-rate.module';
+import { WaterReductionModule } from './water-reduction/water-reduction.module';
 
 @NgModule({
   declarations: [WasteWaterListComponent],
@@ -10,7 +11,8 @@ import { O2UtilizationRateModule } from './o2-utilization-rate/o2-utilization-ra
   imports: [
     CommonModule,
     RouterModule,
-    O2UtilizationRateModule
+    O2UtilizationRateModule,
+    WaterReductionModule
   ]
 })
 export class WasteWaterModule { }
