+++ src/app/calculator/waste-water/water-reduction/water-reduction-results/water-reduction-results.component.css
