+++ src/app/calculator/waste-water/water-reduction/water-reduction.component.spec.ts
