+++ src/app/treasure-hunt/treasure-hunt-report/opportunity-summary.service.ts
@@ -10,7 +10,7 @@ import { MotorDriveService } from '../../calculator/motors/motor-drive/motor-dri
 import { ElectricityReductionService } from '../../calculator/utilities/electricity-reduction/electricity-reduction.service';
 import { ElectricityReductionResults, NaturalGasReductionResults, CompressedAirReductionResults, WaterReductionResults, CompressedAirPressureReductionResults, SteamReductionResults, PipeInsulationReductionResults, TankInsulationReductionResults, AirLeakSurveyOutput } from '../../shared/models/standalone';
 import { NaturalGasReductionService } from '../../calculator/utilities/natural-gas-reduction/natural-gas-reduction.service';
-import { WaterReductionService } from '../../calculator/utilities/water-reduction/water-reduction.service';
+import { WaterReductionService } from '../../calculator/waste-water/water-reduction/water-reduction.service';
 import { PipeInsulationReductionService } from '../../calculator/steam/pipe-insulation-reduction/pipe-insulation-reduction.service';
 import { CompressedAirReductionService } from '../../calculator/compressed-air/compressed-air-reduction/compressed-air-reduction.service';
 import { CompressedAirPressureReductionService } from '../../calculator/compressed-air/compressed-air-pressure-reduction/compressed-air-pressure-reduction.service';
